Added loading icon code for asynchronous calls
Any component that wants to make use of loading circle while waiting
for one or several asynchronous calls has to extend the Loading class
and call super(true) in the constructor. When the asynchronous calls
are finished, simply call this.ready() in the component. The html needs
to have an ngSwitch (see artist.component.html for example) to see
whether loading is false or not.